backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
fpdf2
markdown2
python-dotenv
pytest
ollama
pypdf
